separate cppfeature_access and isi from pyfeatures
